feat: peers in UI (again)
Bring back peer list, this time in "header" bar.

To accommodate, refactors a little to create a SoundMachine per-peer,
enabling the removal of the kudgy peer-id-per-sound hacks.

Also cleans up the peer (i think?) on disconnect, hopefully improving the
stability of the app as peers roll in and out over time.